# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel Wallet",
        "description": "API documentation for user management and financial transactions.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Error registering user"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and get JWT token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "JWT token generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZXhwIjoxNjM5ODI0NjYyfQ.-VZH7Dx6qS4BdF_HQyHtI5Ak0yT_eVf6FY5qISnaGo0"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "John Doe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/reversals/transaction": {
            "post": {
                "tags": [
                    "Reversal"
                ],
                "summary": "Reverse a transaction",
                "operationId": "22326ae9954335ccfde7ebcce89ba587",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_id",
                                    "reason"
                                ],
                                "properties": {
                                    "account_id": {
                                        "description": "ID of the account",
                                        "type": "integer",
                                        "example": 72
                                    },
                                    "transaction_id": {
                                        "description": "ID of the transaction to be reversed",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "reason": {
                                        "description": "Reason for the reversal",
                                        "type": "string",
                                        "example": "Sent to the wrong account"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction reversed successfully"
                    },
                    "400": {
                        "description": "Invalid transaction or insufficient balance"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reversals/deposit": {
            "post": {
                "tags": [
                    "Reversal"
                ],
                "summary": "Reverse a deposit",
                "operationId": "9ca4f1b1b57c4e65428d6d3574b2f898",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "deposit_id",
                                    "reason"
                                ],
                                "properties": {
                                    "account_id": {
                                        "description": "ID of the account",
                                        "type": "integer",
                                        "example": 72
                                    },
                                    "transaction_id": {
                                        "description": "ID of the deposit to be reversed",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "reason": {
                                        "description": "Reason for the reversal",
                                        "type": "string",
                                        "example": "Deposit sent to the wrong account"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deposit reversed successfully"
                    },
                    "400": {
                        "description": "Invalid deposit or insufficient balance"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}